requirements.txt
pyhive
sqlparse
faker
configparser>=3.5.0
kafka-python==1.4.3